Moved the hotload into MakeSearch
